requirements.txt
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.5
django-annoying==0.10.3
django-bower==5.2.0
django-geoposition==0.3.0
django-location-field==2.0.3
django-scheduler==0.8.6.1
gunicorn==19.7.1
icalendar==3.11.7
idna==2.6
oauthlib==2.0.4
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.2
PyJWT==1.5.3
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
urllib3==1.22
whitenoise==3.3.1
